Relocated generate_test.go
