--- examples/testdata/commands/generate/devbox.json
@@ -1,9 +1,0 @@
-{
-  "packages": [],
-  "shell": {
-    "init_hook": null
-  },
-  "nixpkgs": {
-    "commit": "af9e00071d0971eb292fd5abef334e66eda3cb69"
-  }
-}
